Refactor event data types to use unsigned integers
Updated various fields across event structures to use `u64` and `u32` instead of `i64` where unsigned integers are more appropriate. This improves consistency and accuracy in handling data values that should not be negative.

+++ src/event/economy/bounty.rs
@@ -7,7 +7,7 @@ pub struct Reward {
     pub faction: String,
 
     #[serde(rename = "Reward")]
-    pub reward: i64,
+    pub reward: u32,
 }
 
 #[derive(Deserialize, Debug, Default, Clone)]
@@ -31,7 +31,7 @@ pub struct Bounty {
     pub target_localised: Option<String>,
 
     #[serde(rename = "TotalReward")]
-    pub total_reward: i64,
+    pub total_reward: u32,
 
     #[serde(rename = "VictimFaction")]
     pub victim_faction: String,
@@ -43,13 +43,13 @@ pub struct PayBounties {
     pub timestamp: String,
 
     #[serde(rename = "Amount")]
-    pub amount: i64,
+    pub amount: u32,
 
     #[serde(rename = "AllFines")]
     pub all_fines: bool,
 
     #[serde(rename = "ShipID")]
-    pub ship_id: i64,
+    pub ship_id: u64,
 
     #[serde(rename = "BrokerPercentage")]
     pub broker_percentage: f64,
