+++ src/event/economy/sell_weapon.rs
@@ -12,14 +12,14 @@ pub struct SellWeapon {
     pub name_localised: String,
 
     #[serde(rename = "Class")]
-    pub class: i64,
+    pub class: u32,
 
     #[serde(rename = "WeaponMods")]
     pub weapon_mods: Vec<String>,
 
     #[serde(rename = "Price")]
-    pub price: i64,
+    pub price: u32,
 
     #[serde(rename = "SuitModuleID")]
-    pub suit_module_id: i64,
+    pub suit_module_id: u64,
 }
