+++ src/event/economy/trade_micro_resources.rs
@@ -13,7 +13,7 @@ pub struct OfferedMicroResource {
     pub category: String,
 
     #[serde(rename = "Count")]
-    pub count: i64,
+    pub count: u32,
 }
 
 #[derive(Deserialize, Debug, Default, Clone)]
@@ -25,17 +25,17 @@ pub struct TradeMicroResources {
     pub offered: Vec<OfferedMicroResource>,
 
     #[serde(rename = "TotalCount")]
-    pub total_count: i64,
+    pub total_count: u32,
 
     #[serde(rename = "Received")]
     pub received: String,
 
     #[serde(rename = "Count")]
-    pub count: i64,
+    pub count: u32,
 
     #[serde(rename = "Category")]
     pub category: String,
 
     #[serde(rename = "MarketID")]
-    pub market_id: i64,
+    pub market_id: u64,
 }
